requirements.txt
pillow==10.4.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
datasets==3.1.0
nltk==3.8.1
